requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord-py-interactions==5.11.0
discord-py-slash-command==4.2.1
discord-typings==0.7.0
discord.py==2.3.2
emoji==2.9.0
Flask==3.0.0
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
yarl==1.9.4